backend/requirements.txt
# Python 3.11 required
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
groq==0.9.0
google-generativeai==0.7.0
openai==1.40.0
httpx==0.27.0
pandas==3.0.1
numpy==2.2.3
websockets==12.0
python-multipart==0.0.9
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
structlog==24.4.0
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0
slowapi==0.1.9
opencv-python-headless==4.11.0.86
